requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dnspython==2.4.0
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.3
Werkzeug==2.3.6
mongoengine==0.27.0

